setuptools: restructure scripts to use entry_points in setup.py
